requirements.txt
addict==2.4.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cycler==0.11.0
fonttools==4.37.1
idna==3.3
importlib-metadata==4.12.0
kiwisolver==1.4.4
Markdown==3.4.1
matplotlib==3.5.3
mmcv-full==1.5.3
model-index==0.1.11
numpy==1.23.2
opencv-python==4.5.5.64
openmim==0.2.1
ordered-set==4.1.0
packaging==21.3
Pillow==9.2.0
prettytable==3.3.0
Pygments==2.13.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
rich==12.5.1
six==1.16.0
tabulate==0.8.10
torch==1.11.0+cu113
torchvision==0.12.0+cu113
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
yapf==0.32.0
zipp==3.8.1
